feat: add assert_success_rate fixture and refactor success rate validation
